backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.13.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.25.2
python-dotenv>=1.0.0
aiosmtplib>=3.0.1
jinja2>=3.1.2
